Add a Custom Renderer for JSON
Since we are using google/jsonapi, the payload we need to send back in
the response is already encoded as JSON. We need to use a custom render
in order to avoid a double encoding. To to that we use a Custom
Render function created using a closure with the JSON payload.